ai_agent_1/requirements.txt
pandas>=2.0
numpy>=1.24
httpx>=0.24
pyyaml>=6.0
python-dotenv>=1.0
pytz>=2023.3
matplotlib>=3.7
pytest>=7.0
